[Rename] `streaming_management` -->  `event_system`
