+++ hatchling/core/chat/message_history.py
@@ -6,7 +6,7 @@ including user messages, assistant responses, and tool interactions.
 
 from typing import List, Dict, Any, Optional
 from hatchling.core.logging.logging_manager import logging_manager
-from hatchling.core.llm.streaming_management import EventSubscriber, StreamEvent, StreamEventType
+from hatchling.core.llm.event_system import EventSubscriber, StreamEvent, StreamEventType
 from hatchling.config.llm_settings import ELLMProvider
 
 from hatchling.core.llm.data_structures import ToolCallParsedResult
