+++ hatchling/ui/cli_event_subscriber.py
@@ -42,8 +42,8 @@ class UIStateManager:
     def set_only(self, flag: UIStateFlags):
         self.flags = flag
 
-from hatchling.core.llm.streaming_management.event_subscriber import EventSubscriber
-from hatchling.core.llm.streaming_management.stream_data import StreamEvent, StreamEventType
+from hatchling.core.llm.event_system.event_subscriber import EventSubscriber
+from hatchling.core.llm.event_system.stream_data import StreamEvent, StreamEventType
 from hatchling.core.logging.logging_manager import logging_manager
 from prompt_toolkit import print_formatted_text as print_pt
