+++ tests/feature_test_ollama_streaming.py
@@ -13,7 +13,7 @@ from hatchling.config.llm_settings import ELLMProvider
 from hatchling.config.settings import AppSettings
 from hatchling.config.ollama_settings import OllamaSettings
 from hatchling.core.llm.providers import ProviderRegistry
-from hatchling.core.llm.streaming_management import (
+from hatchling.core.llm.event_system import (
     EventPublisher,
     ContentPrinterSubscriber,
     UsageStatsSubscriber,
